devide main to files
